Abort HTTP requests after a timeout
If a HTTP request takes longer than the specified timeout,
the request will be aborted and an error will be printed to the console.

Reaching the timeout will result in an error and the operation will be aborted.

Syncing repositories is an exception: the error will just be printed to
the console and not cause an error, because one mirror might be
unreachable doesn't necessarily mean all mirrors are unreachable.

The timeout only affects reading response headers.
The timeout is currently set to 10 seconds.

Signed-off-by: Timo Reichl <thetredev@gmail.com>